requirements.txt
altgraph==0.17.5
asgiref==3.11.1
blinker==1.9.0
bottle==0.13.4
brotli==1.2.0
cffi==2.0.0
chardet==7.4.0.post1
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.9.0
Django==6.0.3
et_xmlfile==2.0.0
Flask==3.0.0
fonttools==4.62.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pefile==2023.2.7
pillow==12.1.1
proxy_tools==0.1.0
psycopg==3.3.3
psycopg-binary==3.3.3
pycparser==3.0
pydyf==0.12.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2026.3
pypdf==6.9.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
reportlab==4.1.0
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
typing_extensions==4.15.0
tzdata==2025.3
waitress==3.0.2
weasyprint==68.1
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.11.0
zopfli==0.4.1
